RIDs are now stored as BIGINT instead of INT_MAX

+++ Application/database/migrations/2015_11_13_150427_createTrainTimesTable.php
@@ -19,7 +19,7 @@ class CreateTrainTimesTable extends Migration
             $table->dateTime('from_time');
             $table->string('to_tpl');
             $table->dateTime('to_time');
-            $table->integer( 'rid' );
+            $table->bigInteger( 'rid' );
         });
     }
